requirements.txt
fastapi==0.115.8
google-genai==1.2.0
openai==1.63.0
PyMuPDF==1.25.3
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
uvicorn==0.34.0
openpyxl==3.1.5

